+++ ddipy/cli.py
@@ -1,46 +1,43 @@
 import click
-from omicsdi_client import OmcicsClient, url_path_join
 import os
 from urllib.parse import urlsplit
+import json
+import requests
+from urllib3.util.retry import Retry
+import urllib.request
+import ftplib
+import re
 import shutil
+import click
 import sys
+from ddipy.dataset_client import DatasetClient
 
-client = OmcicsClient()
 
+def url_path_join(*args):
+    """Join path(s) in URL using slashes"""
 
-def make_dir(output, acc_number):
-
-    if output:
-        dir_path = url_path_join(output, acc_number)
-    else:
-        dir_path = acc_number
-
-    if os.path.exists(dir_path):
-        shutil.rmtree(dir_path)
-
-    os.makedirs(dir_path)
-
-    return dir_path
-
+    return '/'.join(s.strip('/') for s in args)
 
 def source_from_id(acc):
-    api_output = client.get_source(acc)
-    if api_output['datasets']:
-        if len(api_output['datasets']) > 1:
+    client = DatasetClient()
+    api_output = client.search(acc, "id")
+    if api_output.datasets:
+        datasets = api_output.datasets
+        if len(datasets) > 1:
             hits = []
-            for dataset in api_output['datasets']:
-                if dataset['id'] == acc:
+            for dataset in datasets:
+                if dataset.accession == acc:
                     hits.append(dataset)
 
             if len(hits) == 1:
-                source = hits[0]['source']
+                source = hits[0].database
                 return source
             else:
                 click.echo('INFO')
                 click.echo(f'{acc} is not a unique ID, leading to multiple hits.')
                 sys.exit()
         else:
-            source = api_output['datasets'][0]['source']
+            source = datasets[0].database
             return source
     else:
         click.echo('INFO')
@@ -49,10 +46,11 @@ def source_from_id(acc):
 
 
 def file_links(source, acc):
-    api_output = client.get_data(source, acc)
+    client = DatasetClient()
+    api_output = client.get_dataset_details(source, acc)
     file_links = []
-    if api_output['file_versions']:
-        files = api_output['file_versions'][0]['files']
+    if api_output.files:
+        files = api_output.files[0]['files']
         for ext, files in files.items():
             for url in files:
                 file_links.append(url)
@@ -94,6 +92,48 @@ def print_version(ctx, param, value):
     click.echo('Version 0.1')
     ctx.exit()
 
+def download_ftp_files( domain, project_dir, dir_path, filename):
+    """Download ftp files in given directory"""
+
+    with ftplib.FTP(domain) as ftp:
+        try:
+            ftp.login()
+            ftp.cwd(project_dir)
+            file_path = url_path_join(dir_path, filename)
+            localfile = open(file_path, 'wb')
+            ftp.retrbinary(
+                "RETR " + filename, localfile.write)
+            localfile.close()
+
+        except ftplib.all_errors as e:
+            click.echo('--> FTP error:', e)
+            click.echo('--> Please check if ' + filename +
+                    ' exists in ' + project_dir + ' on the domain ' + domain)
+
+def download_http_files(self, file_url, filename, dir_path):
+    """Download http files in given directory"""
+
+    file_path = url_path_join(dir_path, filename)
+    try:
+        session = requests.Session()
+        retry = Retry(connect=3, backoff_factor=15)
+        adapter = requests.adapters.HTTPAdapter(max_retries=retry)
+        session.mount('http://', adapter)
+        session.mount('https://', adapter)
+        r = session.get(file_url)
+
+        if r.status_code != requests.codes.ok:
+            click.echo('--> ERROR Response ' + str(r.status_code) + ', reason: ' + str(r.reason))
+            click.echo('--> Please check if ' + file_url + ' is reachable')
+
+        else:
+            with urllib.request.urlopen(file_url) as response, open(file_path, 'wb') as out_file:
+                shutil.copyfileobj(response, out_file)
+
+    except urllib.request.HTTPError as e:
+        click.echo('-->', e)
+        click.echo('--> Please check if ' + file_url + ' is reachable')
+
 
 @click.command(context_settings={'help_option_names': ['-h', '--help']})
 @click.option('--version', '-v',  callback=print_version, expose_value=False, is_eager=True, is_flag=True, help="Print version number")
@@ -103,17 +143,11 @@ def print_version(ctx, param, value):
     help='Use this flag to download the files in the current directory or a specified output directory',
 )
 @click.option(
-    '--output', '-o',  default=None,  type=click.Path(exists=True),
+    '--output', '-o',  default=".",  type=click.Path(exists=True),
     help='Output directory when downloading files (default: CWD)',
 )
 def main(acc_number, download, output):
-    """
-\b                                                     
-   ___        _       ___  _    __     _      _            
-  / _ \ _ __ (_)__ __|   \(_)  / _|___| |_ __| |_  ___ _ _ 
- | (_) | '  \| / _(_-< |) | | |  _/ -_)  _/ _| ' \/ -_) '_|
-  \___/|_|_|_|_\__/__/___/|_| |_| \___|\__\__|_||_\___|_|                                                                                            
-\b                                                   
+    """                                            
     Command Line Interface to fetch data from OmicsDi.
     """
 
@@ -122,17 +156,16 @@ def main(acc_number, download, output):
 
     # Download section
     if download and file_urls:
-        dir_path = make_dir(output, acc_number)
         for file_url in file_urls:
             info = url_info(file_url)
             click.echo("Downloading...  " + filename_process(info['filename']))
             if info['scheme'] == 'ftp':
-                client.download_ftp_files(
-                    info['domain'], info['project_dir'], dir_path, info['filename'])
+                download_ftp_files(
+                    info['domain'], info['project_dir'], output, info['filename'])
             elif info['scheme'] == 'https' or info['scheme'] == 'http':
                 info = url_info(file_url)
-                client.download_http_files(
-                    file_url, filename_process(info['filename']), dir_path)
+                download_http_files(
+                    file_url, filename_process(info['filename']), output)
             else:
                 click.echo('--> Scheme is not supported for ' + file_url)
     # DEFAULT: Printing URLS when -d is not given
@@ -142,4 +175,4 @@ def main(acc_number, download, output):
 
 
 if __name__ == "__main__":
-    main()
+    main(sys.argv[1:])
